Conflicts:
	protected/components/ActiveForm.php
	protected/components/EntryButtonColumn.php
	protected/components/SecurityManager.php
	protected/components/WebUser.php

+++ protected/models/Entry.php
@@ -15,6 +15,9 @@
  * @property integer $userId
  * @property string  $username
  * @property int     $viewCount
+ *
+ * @property Category[] $categories
+ * @property int[] $categoryIds
  */
 class Entry extends CActiveRecord
 {
@@ -26,15 +29,17 @@ class Entry extends CActiveRecord
     public function attributeLabels()
     {
         return array(
-            'comment'   => 'Comment',
-            'id'        => 'ID',
-            'name'      => 'Name',
-            'password'  => 'Password',
-            'tagList'   => 'Tags',
-            'url'       => 'URL',
-            'userId'    => 'User',
-            'username'  => 'Username',
-            'viewCount' => 'View Count'
+            'comment'     => 'Comment',
+            'categories'  => 'Categories',
+            'categoryIds' => 'Categories',
+            'id'          => 'ID',
+            'name'        => 'Name',
+            'password'    => 'Password',
+            'tagList'     => 'Tags',
+            'url'         => 'URL',
+            'userId'      => 'User',
+            'username'    => 'Username',
+            'viewCount'   => 'View Count'
         );
     }
 
@@ -46,6 +51,7 @@ class Entry extends CActiveRecord
     public function afterDelete()
     {
         $this->deleteTags();
+        $this->deleteCategories();
         return parent::afterDelete();
     }
 
@@ -56,32 +62,49 @@ class Entry extends CActiveRecord
      */
     public function afterSave()
     {
-        // delete all tag relations
-        $this->deleteTags();
-
-        // save tags and tag relations
-        foreach ($this->tags as $tag)
+        // after update & create
+        if (in_array($this->scenario, array('update', 'create')))
         {
-            // try to receive tag from db
-            $model = Tag::model()->name( $tag->name )->userId( Yii::app()->user->id )->find();
+            // delete all tag relations
+            $this->deleteTags();
 
-            if (!is_object($model))
+            // save tags and tag relations
+            foreach ($this->tags as $tag)
             {
-                $model = $tag;
+                // try to receive tag from db
+                $model = Tag::model()->name( $tag->name )->userId( Yii::app()->user->id )->find();
+
+                if (!is_object($model))
+                {
+                    $model = $tag;
+                }
+
+                // save tag
+                $model->name = $tag->name;
+                $model->save();
+
+                // save relation
+                $relation = new EntryHasTag();
+                $relation->entryId = $this->id;
+                $relation->tagId   = $model->id;
+                $relation->save();
             }
 
-            // save tag
-            $model->name = $tag->name;
-            $model->save();
+            // delete all categories
+            $this->deleteCategories();
 
-            // save relation
-            $relation = new EntryHasTag();
-            $relation->entryId = $this->id;
-            $relation->tagId   = $model->id;
-            $relation->save();
+            // save categories
+            foreach ($this->categories as $category)
+            {
+                /* @var Category $category */
+                $relation = new CategoryHasEntry();
+                $relation->entryId    = $this->id;
+                $relation->categoryId = $category->id;
+                $relation->save();
+            }
+
+            return parent::afterSave();
         }
-
-        return parent::afterSave();
     }
 
 
@@ -102,17 +125,53 @@ class Entry extends CActiveRecord
 
 
     /**
+     * @return void
+     */
+    public function deleteCategories()
+    {
+        // runs only after delete and update
+        if (in_array($this->scenario, array('update', 'delete')))
+        {
+            $relations = CategoryHasEntry::model()->entryId($this->id)->findAll();
+
+            foreach ($relations as $relation)
+            {
+                /* @var CategoryHasEntry $relation */
+                $relation->delete();
+            }
+        }
+    }
+
+
+    /**
      *
      * @return void
      */
     public function deleteTags()
     {
-        $relations = EntryHasTag::model()->entryId($this->id)->findAll();
-
-        foreach ($relations as $relation)
+        // runs only after delete and update
+        if (in_array($this->scenario, array('update', 'delete')))
         {
-            $relation->delete();
+            $relations = EntryHasTag::model()->entryId($this->id)->findAll();
+
+            foreach ($relations as $relation)
+            {
+                $relation->delete();
+            }
         }
+    }
+
+
+    public function getCategoryIds()
+    {
+        $ids = array();
+
+        foreach ($this->categories as $category)
+        {
+            $ids[] = $category->id;
+        }
+
+        return $ids;
     }
 
 
@@ -195,8 +254,9 @@ class Entry extends CActiveRecord
     public function relations()
     {
         return array(
-            'user' => array(self::BELONGS_TO, 'User', 'userId'),
-            'tags' => array(self::MANY_MANY, 'Tag', 'EntryHasTag(entryId, tagId)'),
+            'user'       => array(self::BELONGS_TO, 'User', 'userId'),
+            'tags'       => array(self::MANY_MANY, 'Tag', 'EntryHasTag(entryId, tagId)'),
+            'categories' => array(self::MANY_MANY, 'Category', 'category_has_entry(entryId, categoryId)'),
         );
     }
 
@@ -209,6 +269,8 @@ class Entry extends CActiveRecord
     {
         return array(
             array('comment', 'default', 'value' => NULL),
+
+            array('categoryIds', 'safe'),
 
             array('id', 'safe', 'on'=>'search'),
 
@@ -243,6 +305,7 @@ class Entry extends CActiveRecord
     public function search()
     {
         $criteria = new CDbCriteria();
+        $alias    = $this->getTableAlias();
 
         // by search term
         if (Yii::app()->request->getParam('q') != null)
@@ -250,30 +313,42 @@ class Entry extends CActiveRecord
             $term = Yii::app()->request->getParam('q');
 
             $criteria->distinct = true;
-            $criteria->join = 'LEFT JOIN EntryHasTag AS eht ON eht.entryId=t.id '
+            $criteria->join = 'LEFT JOIN EntryHasTag AS eht ON eht.entryId=' . $alias  .'.id '
+                            . 'LEFT JOIN category_has_entry AS che ON che.entryId=' . $alias  .'.id '
+                            . 'LEFT JOIN Category ON Category.id=che.categoryId '
                             . 'LEFT JOIN Tag ON Tag.id=eht.tagId';
 
-            $criteria->compare('t.name', $term, true, 'OR');
-            $criteria->compare('t.url', $term, true, 'OR');
-            $criteria->compare('t.comment', $term, true, 'OR');
-            $criteria->compare('t.username', $term, true, 'OR');
+            $criteria->compare($alias . '.name', $term, true, 'OR');
+            $criteria->compare($alias . '.url', $term, true, 'OR');
+            $criteria->compare($alias . '.comment', $term, true, 'OR');
+            $criteria->compare($alias . '.username', $term, true, 'OR');
             $criteria->compare('Tag.name', $term, true, 'OR');
+            $criteria->compare('Category.name', $term, true, 'OR');
         }
 
         // by detail search
         else
         {
-            $criteria->compare('t.name', $this->name, true);
-            $criteria->compare('t.url', $this->url, true);
-            $criteria->compare('t.comment', $this->comment);
-            $criteria->compare('t.username', $this->username);
+            $criteria->compare($alias . '.name', $this->name, true);
+            $criteria->compare($alias . '.url', $this->url, true);
+            $criteria->compare($alias . '.comment', $this->comment);
+            $criteria->compare($alias . '.username', $this->username);
 
             if (strlen($this->tagList) > 0)
             {
                 $c = new CDbCriteria();
-                $c->join = 'INNER JOIN EntryHasTag AS eht ON eht.entryId=t.id '
+                $c->join = 'INNER JOIN EntryHasTag AS eht ON eht.entryId=' . $alias . '.id '
                          . 'INNER JOIN Tag ON Tag.id=eht.tagId';
                 $c->compare('Tag.name', $this->tagList);
+                $criteria->mergeWith($c);
+            }
+
+            if (count($this->categoryIds) > 0)
+            {
+                $c = new CDbCriteria();
+                $c->join = 'INNER JOIN category_has_entry AS che ON che.entryId=' . $alias . '.id '
+                    . 'INNER JOIN Category ON Category.id=che.categoryId';
+                $c->addInCondition('Category.id', $this->categoryIds);
                 $criteria->mergeWith($c);
             }
         }
@@ -298,6 +373,26 @@ class Entry extends CActiveRecord
         else
         {
             $this->encryptedPassword = '';
+        }
+    }
+
+
+    /**
+     * @param int[] $v
+     */
+    public function setCategoryIds($v)
+    {
+        if (is_array($v))
+        {
+            foreach ($v as $id)
+            {
+                $category = Category::model()->findByPk($id);
+
+                if ($category instanceof Category)
+                {
+                    $this->categories = array_merge($this->categories, array($category));
+                }
+            }
         }
     }
 
@@ -333,8 +428,12 @@ class Entry extends CActiveRecord
      */
     public function incrementViewCounter()
     {
+        $oldScenario = $this->scenario;
+        $this->scenario = 'incrementCounter';
         $this->viewCount++;
         $this->save(true, array('viewCount'));
+        $this->scenario = $oldScenario;
+
     }
 
 }
